+++ src/main/java/lotto/model/domain/CompareResult.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.domain;
 
 public class CompareResult {
