+++ src/main/java/lotto/model/domain/Lotto.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.domain;
 
 import java.util.List;
 import java.util.Set;
