+++ src/main/java/lotto/model/service/LottoPublisher.java
@@ -1,11 +1,14 @@
-package lotto.model;
+package lotto.model.service;
 
-import static lotto.model.LottoConstant.MAX_NUMBER;
-import static lotto.model.LottoConstant.MIN_NUMBER;
+import static lotto.model.constant.LottoConstant.MAX_NUMBER;
+import static lotto.model.constant.LottoConstant.MIN_NUMBER;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.ArrayList;
 import java.util.List;
+import lotto.model.domain.Lotto;
+import lotto.model.domain.PublishedLotto;
+import lotto.model.domain.PurchasePrice;
 
 public class LottoPublisher {
