+++ src/main/java/lotto/model/service/LottoService.java
@@ -1,6 +1,9 @@
-package lotto.model;
+package lotto.model.service;
 
 import java.util.List;
+import lotto.model.domain.CompareResult;
+import lotto.model.domain.PublishedLotto;
+import lotto.model.domain.WinningLotto;
 
 public class LottoService {
